Move header to include folder to make it more consumable for other projects
