+++ tests/test.cc
@@ -1,6 +1,6 @@
 #define CATCH_CONFIG_MAIN  // This tells Catch to provide a main() - only do this in one cpp file
 #include "catch.hpp"
-#include "../float16_t.hpp"
+#include "../include/float16_t.hpp"
 #include <cmath>
 #include <iostream>
 #include <bitset>
